% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dat.R
\name{mod_dat_country_ui}
\alias{mod_dat_country_ui}
\title{mod_dat.R}
\usage{
mod_dat_country_ui(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
