% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dots.R
\name{mod_dots_country_ui}
\alias{mod_dots_country_ui}
\title{mod_dots.R}
\usage{
mod_dots_country_ui(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
