% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pop_mean.R
\name{mod_pop_mean_ui}
\alias{mod_pop_mean_ui}
\alias{mod_pop_mean_server}
\title{mod_pop_mean_ui and mod_pop_mean_server}
\usage{
mod_pop_mean_ui(id)

mod_pop_mean_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
