% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holder.R, R/mod_recent_sub.R
\name{mod_recent_mean_sub_ui}
\alias{mod_recent_mean_sub_ui}
\title{mod_recent_sub_value}
\usage{
mod_recent_mean_sub_ui(id)

mod_recent_mean_sub_ui(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.

A shiny Module.
}
\keyword{internal}
