% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_recent.R
\name{mod_recent_mean_ui}
\alias{mod_recent_mean_ui}
\title{mod_recent_value}
\usage{
mod_recent_mean_ui(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
