% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rural.R
\name{mod_rural_ui}
\alias{mod_rural_ui}
\title{mod_recent_sub.R}
\usage{
mod_rural_ui(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
