% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_social.R
\name{mod_social_ui}
\alias{mod_social_ui}
\title{mod_social_ui and mod_social_server}
\usage{
mod_social_ui(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
Make the social section
}
\keyword{internal}
