% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trends.R
\name{mod_trends_mean_ui}
\alias{mod_trends_mean_ui}
\title{mod_trens.R}
\usage{
mod_trends_mean_ui(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
