% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{theme_gdocs}
\alias{theme_gdocs}
\title{Theme based on goolge docs}
\usage{
theme_gdocs(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Theme similar to the default look of charts in Google Docs.
}
