% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme_new.R
\name{theme_hefpi}
\alias{theme_hefpi}
\title{Theme based on goolge docs}
\usage{
theme_hefpi(
  base_size = 12,
  base_family = "Helveticaserif",
  plot_title_size = 15,
  plot_title_color = "#757575",
  sub_title_size = 10,
  sub_title_color = "#757575",
  axis_title_size = 1,
  axis_title_color = "#757575",
  x_axis_color = "#5A5A5A",
  x_axis_size = 12,
  x_axis_angle = 0,
  x_axis_hjust = 0,
  x_axis_vjust = 0,
  y_axis_color = "#5A5A5A",
  y_axis_size = 12,
  y_axis_angle = 0,
  y_axis_hjust = 0,
  y_axis_vjust = 0,
  grid_major_x = "#cccccc",
  grid_minor_x = NA,
  grid_major_y = "#cccccc",
  grid_minor_y = NA,
  x_axis_line = "#cccccc",
  y_axis_line = "#cccccc",
  legend_position = "right",
  legend_direction = "vertical",
  legend_text_size = 1,
  legend_text_color = "#757575"
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Theme similar to the default look of charts in Google Docs.
}
